Android: small fix for external video surface on 64-bit.
BUG=346626, 354405

Review URL: https://codereview.chromium.org/268653002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@267552 0039d316-1c4b-4281-b951-d872f2087c98

+++ android_webview/java/src/org/chromium/android_webview/ExternalVideoSurfaceContainer.java
@@ -64,7 +64,7 @@ public class ExternalVideoSurfaceContainer implements SurfaceHolder.Callback {
     private static WeakReference<ExternalVideoSurfaceContainer> sActiveContainer =
             new WeakReference<ExternalVideoSurfaceContainer>(null);
 
-    private final int mNativeExternalVideoSurfaceContainer;
+    private final long mNativeExternalVideoSurfaceContainer;
     private final ContentViewCore mContentViewCore;
     private int mPlayerId = INVALID_PLAYER_ID;
     private SurfaceView mSurfaceView;
@@ -86,7 +86,7 @@ public class ExternalVideoSurfaceContainer implements SurfaceHolder.Callback {
      */
     public static class Factory {
         public ExternalVideoSurfaceContainer create(
-                int nativeExternalVideoSurfaceContainer, ContentViewCore contentViewCore) {
+                long nativeExternalVideoSurfaceContainer, ContentViewCore contentViewCore) {
             return new ExternalVideoSurfaceContainer(
                     nativeExternalVideoSurfaceContainer, contentViewCore);
         }
@@ -100,12 +100,12 @@ public class ExternalVideoSurfaceContainer implements SurfaceHolder.Callback {
 
     @CalledByNative
     private static ExternalVideoSurfaceContainer create(
-            int nativeExternalVideoSurfaceContainer, ContentViewCore contentViewCore) {
+            long nativeExternalVideoSurfaceContainer, ContentViewCore contentViewCore) {
         return sFactory.create(nativeExternalVideoSurfaceContainer, contentViewCore);
     }
 
     protected ExternalVideoSurfaceContainer(
-            int nativeExternalVideoSurfaceContainer, ContentViewCore contentViewCore) {
+            long nativeExternalVideoSurfaceContainer, ContentViewCore contentViewCore) {
         assert contentViewCore != null;
         mNativeExternalVideoSurfaceContainer = nativeExternalVideoSurfaceContainer;
         mContentViewCore = contentViewCore;
